requirements.txt
beautifulsoup4==4.14.3
Flask-Limiter==4.1.1
Flask==3.1.2
gunicorn==23.0.0
pillow==12.0.0
python-dotenv==1.2.1
requests==2.32.5
Werkzeug==3.1.4
